[FEATURE] Add Package Filtering by Path and Type
This makes getting a particular kind of package, by path or type, or
first class Flow feature. The docs say that you can put packages in any
subdirectory of Packages/, but if you want to get a list of packages in
any one of those directories, you had to do your own filter the list of
packages yourself.

For example:

Neos can use this to get typo3-flow-site packages in Packages/Sites/.
Builders can use this to get a list of typo3-flow-boilerplate packages.

This change is backwards compatible since it doesn't change any
behavior, it just adds some additional methods to the PackageManager.

Change-Id: Ic87ebaece612e25898318795c748941e5a96b8cb
Releases: master