[FEATURE] Catch and log exceptions occurring in echoExceptionWeb
If, during the rather complex process of custom exception renderings,
again an exception occurs, this exception is again caught silently
and logged. In such a case there won't be anything output (white page),
but still a 500 Internal Server Error is sent, and the actual reason
for the exception can still be found in the log.

Change-Id: I8f9827b6ccf695672d7c80186ea5efe1ec0db4b4
Releases: master