[FEATURE] Let cache:warmup warm up every configured configuration type
When having done a flow:cache:warmup, the Configuration Manager only cached
the configurations which were necessary during the CLI command invocation,
so, for example, Routes, Views, NodeTypes etc. were not cached.

This change introduces a SignalSlot listening to the warmupCache signal,
and touching every configured configuration type in order to make it
present in the configurations array, ready for later storing.

Change-Id: I82f508b977cc4146fa5fd46ab82581a5f4156ea5
Resolves: #51652
Releases: master