[BUGFIX] Published resources don't support symlinks
Fixes an issue with the publication of persistent resources which
resulted in inaccessible resources after a few deployments with
TYPO3.Surf.

TYPO3.Surf heavily relies on symbolic links for pointing to the
currently active release. The key problem was that the symbolic
link which is put into the Web/_Resources folder used an absolute
path (realpath) to the resource file in Data/Persistent/ rather than
a possible symbolic link. This worked fine as long as the actual
directory (previous release) exists, but breaks as soon as old
releases have been removed.

Resolves: #51676
Releases: 2.0, master
Change-Id: I5585a23eb8db414cbce8f181bf7d73d5c41e0bb9