[TASK] Resources are published with relative symlinks
Fixes an issue with the publication of persistent resources which
used absolute paths rather than relative paths for generating symbolic
links.

TYPO3.Surf heavily relies on symbolic links for pointing to the
currently active release. The key problem was that the symbolic
link which is put into the Web/_Resources folder used an absolute
path (realpath) to the resource file in Data/Persistent/ rather than
a possible symbolic link. This worked fine as long as the actual
directory (previous release) exists, but breaks as soon as old
releases have been removed.

Resolves: #51676
Resolves: #51809
Releases: master
Change-Id: I2e84cd8f9abd41578a87791813d03e764528feb9