[!!!][TASK] Refactor and clarify the database exception handling
This change differentiates between connection errors and missing
tables in the database which were before handled equally.
It also writes the full exception to the log for debugging.

The former DatabaseConnectionException has been moved to
Exception\DatabaseConnectionException, a code migration is
provided with the change.

Change-Id: I048360fc8883942d6561be92c2d0ad915e9f164e
Releases: master
Migration: TYPO3.Flow-201310031523