[BUGFIX] Object Route Parts fall back to technical identifier
As described in the routing section of the documentation, the
`Object Route Parts`` should use the technical identifier to
represent an object if no uriPattern is specified and the
specified object does not contain any identity properties.

Instead an exception was thrown if no uriPattern could be
determined because ``ClassSchema::getIdentityProperties()``
doesn't return the property name of the technical identifier.

This change fixes this by falling back by skipping the
ObjectPathMapping if no uriPattern is defined.

Change-Id: Ic06641be676ab514b8004164e90ffd15b6730cca
Releases: master, 2.0
Fixes: #51082