[TASK] Get rid of RouterCachingAspect
This moves code from the RouterCache AOP aspect to the Router itself
making the aspect redundant.

Before the cache handling of the Routing Framework has been extracted
to an AOP aspect in order to keep the Router code simple. Since the
introduction of the RouterCachingService (#52452) this is no longer
required.

Change-Id: I43fc35c55fdb69e1e67e08782ab0928bc37218de
Resolves: #52704
Related: #52452
Releases: master