[BUGFIX] Role handling depends on an active database connection
The PolicyService now depends on having an active database connection.
This connection is not always available (like in the Neos setup) and
is not required at all for authentication using in memory roles like
the system roles.

This change updates the PolicyService to ignore synchronization if
the RoleRepository has no active database connection, and makes the
FileBasedSimpleKeyProvider use a Role object which only exists in
memory.

Releases: master
Change-Id: I83f8904cf3ecd086eb86f46e01dd602f592c8e19