[BUGFIX] REMOTE_AUTHORIZATION strips "Basic" string
Currently the .htaccess file passes the "Authorization" header
to the REMOTE_AUTHORIZATION environment variable by stripping
the "Basic" string. This means as soon as the authentication
token tries to authenticate, the only remaining thing is the
base64 encoded value of username/password without the "Basic"
indicator. This leads to no basic auth possible when running
CGI.

Change-Id: Ia316a732c1eecade595f67928fc1ce6935731af0
Resolves: #53010
Releases: master, 2.2, 2.1