[BUGFIX] var_dump() is unstyled if Flow is not in docroot
If TYPO3 Flow is not directly in the document root the var_dump
output is unstyled as the path to the external CSS can not be
loaded. By loading all CSS inline and replacing the background
image by CSS there's no need to load external resources anymore.

Change-Id: I674363b0ec8ef20b1787b9e67acfdf520d90d474
Related: #52711
Releases: master