[!!!][FEATURE] Deprecated base tag
Currently the ``UriBuilder`` creates relative URLs with relative paths
in the form "some/path.html" by default.
This change adjusts the ``UriBuilder`` to prefix URLs with the scripts
request path (usually "/" or "/document-root/").

This makes the ``<base />`` tag redundant for most scenarios, even
when working with relative URLs.

This is not a breaking change per se, because using the base tag
doesn't break the behavior. But it has an impact on all relative URLs
created by the ``UriBuilder``.

If you expect the rendered HTML to contain URLs in the form
"some/path.html" (that is: without leading slash) you might need to
adjust your code or set the compatibilty flag that is introduced with
this change::

 TYPO3:
   Flow:
     compatibility:
       uriBuilder:
         createRelativePaths: TRUE

NOTE: This flag will be deprecated with version 2.2 and removed with
version 3.0 of Flow. It's recommended to adjust your code so that it
does not rely on the generated URLs

Change-Id: I2d68e4b22a23afebe0c9dfcf55342b2304e9be9a
Resolves: #53827
Releases: master