[BUGFIX] Persisting of objects as serialized data not supported
The ``ReflectionService`` does not add properties to metadata information
of classes if they are not tagged with ``@Flow\Entity`` or
``@Flow\ValueObject``.
This makes it f.e. impossible to store an object as serialized object.
Previously any object (except some whitelisted) needed at least a
``@ORM\OneToOne`` annotation to be persisted in a property.

To be able to store f.e. ``UserSettings``, ``BasketItems``, ``Sessions``
or DTOs the doctrine Annotation ``@ORM\Column(type="object")`` is now
supported.

The whitelist was removed from ``ReflectionService`` in order to allow
any class to define custom column types.

The ``ClassSchema`` does no longer contain information about
properties declared ``@Flow\Transient``, the same is true for properties
annotated with ``@Flow\Inject`` so that injected properties are never
persisted.

If an object has a property which holds an object and that property is
annotated with ``@ORM\Column(type="object")` that will be stored as 
expected (i.e. as a serialized PHP object). If no such annotation is given
and no other relation type annotation is given, an exception is thrown as
before.

Working Example::

  /**
  * @var \Acme\Foo\Bar;
  * @ORM\Column(type="object")
  */
  protected $bar;

In the example above the doctrine behavior for object types matches and
invokes ``\Doctrine\DBAL\Types\ObjectType``

Also added some tests for other extended types supported by Doctrine:

* \Doctrine\DBAL\Types\Type::SIMPLE_ARRAY
* \Doctrine\DBAL\Types\Type::JSON_ARRAY
* \Doctrine\DBAL\Types\Type::DATETIME
* \Doctrine\DBAL\Types\Type::DATETIMETZ
* \Doctrine\DBAL\Types\Type::DATE
* \Doctrine\DBAL\Types\Type::TIME

Please note:

- DATETIMETZ is not supported by several dbs (i.e. mysql):
  @see http://doctrine-dbal.readthedocs.org/en/latest/reference/known-vendor-issues.html#datetimetz
  If you want to handle different timezone informations in mysql:
  @see http://doctrine-orm.readthedocs.org/en/latest/cookbook/working-with-datetime.html#handling-different-timezones-with-the datetime-type
- DATETIME does not consider timezones and always applies the servers
  date.timezone:
  @see http://doctrine-orm.readthedocs.org/en/latest/cookbook/working-with-datetime.html#default-timezone-gotcha
- SIMPLE_ARRAY and JSON_ARRAY will always be hydrated as array, they
  will never be loaded as NULL, even if marked nullable

Change-Id: I6d759cffd9a46cbd789585d92324f58e66d3e9ae
Fixes: FLOW-254
Releases: master, 3.0