[BUGFIX] Allow uppercase rewrite regex expression for session id
The usage of the SecurityPublishingConfiguration in combination with
the fluid resource viewhelper generates a link which includes the 
users sessions id. This session id includes uppercase characters. 
The mod_rewrite rule for this private resources doesnt allow uppercase
characters for the session id, so the user gets a 404.

Fixes: #54973
Releases: master

Change-Id: I39e9b1922f5ef052b040c04467ea0c041aaec698