[BUGFIX] Make Doctrine service return correct packages for migrations
With a previous fix ``PackageManager::getPackageOfObject()`` now
compares the namespace of a given object rather than it's location
on the disk.
This broke the behavior of
``Doctrine\Service::getPackageKeyFromMigrationVersion()`` because
doctrine migrations all have the same namespace ``\TYPO3\Flow\..``

With this fix the doctrine service compares the file paths again.

Change-Id: Ibaeea8993644a09c7266ff3725d9912e1b91d8dd
Related: #55309
Releases: master