[!!!][FEATURE] Use PHP YAML extension if available
If you have the PECL YAML extension installed it will be used
to parse your configuration files, which results in a tremendous
speed improvement, especially in Development context.

As this YAML parser is stricter also all yaml files are adapted.

This is breaking if the YAML extension is installed and you
have invalid YAML files.

Change-Id: Ibcec9ed8dc0b06830163ecee0b1648c73f988440
Releases: master