[TASK] Optimize arrayMergeRecursiveOverrule by removing recursion
Method calls in PHP are generally slow. As we rely heavily on this
function in several parts it's worthwile to micro-optimize the
implementation to get rid of recursion. Running the unit tests showed
a performance improvement of 25% by using a stack based approach instead
of nested calls. The effect could be even higher for highly nested
arrays (like configuration).

Change-Id: I165c2c0297d1ecbda7dc8ef42820b3cbbe934e2c
Releases: master