[!!!][TASK] Use JSON instead of serialize() in FileMonitor
For larger arrays JSON is faster than serialize() so the FileMonitor
that caches large associative arrays is switched to use json_encode()
and json_decode() and a StringBackend instead of the VariableBackend
that serializes.

This could be breaking if you implemented your own Strategy and use the
Flow_Monitor cache there because it will no longer accept all kinds of
variables but only strings.

Change-Id: I2541ac1671f6559f6c4c93d55452b8d8c3065888
Releases: master