[FEATURE] Convenience method to create custom FileMonitors
Exposes a static method in the FileMonitor class to create
a new FileMonitor instance during boot time in order to use
the FileMonitor for custom purposes.

Additionally add a new method to monitor a directory with a given
filename pattern and fix a bug where the removal of subsequently created
files was not tracked.

Change-Id: Id4aebb951f27b7befe1c1be81301b12679440bfc
Releases: master, 2.2