[!!!][TASK] Decoupling of TYPO3.Party package
This makes the use of the ``TYPO3.Party`` package optional, hence fully
replaceable. The Account does not have a tight coupling to the
``AbstractParty`` Aggregate Root anymore, so complete customized
solutions are usable now.

This change deprecates the following methods:

* ``Account::getParty()``
* ``Account::setParty()``
* ``Security\Context::getParty()``
* ``Security\Context::getPartyByType()``

They still work if the party package is installed, but usage of their
methods should be replaced with custom service calls (see party package
for an example of a simple ``PartyService``).

This is a breaking change because it removes the dependency to the
``TYPO3.Party`` package. If a package makes use of that package (e.g. by
extending the ``AbstractParty`` model, an entry::

  "typo3/party": "~3.0"

in the "require" section of the ``composer.json`` file is to be added!
Besides this change adjusts the ``getParty()`` and ``setParty()``
methods of ``Account`` and rearranges the database structure; so in
cases where these changes might influence userland code, adjustments
might be necessary.

Change-Id: I1a5dee5729fa73229b29d11c9f828eb803eb6b24
Fixes: FLOW-5
Releases: master