[BUGFIX] Re-add styling for standard tags in CLI
With the integration of ``symfony/console`` in #49016 the support for
some HTML tags was lost, namely ``<b>``, ``<i>``, ``<u>``, ``<em>``
and ``<strike>``.

This change registers these tags again in the default ConsoleOutput.

Note: In order to get the formatted output to work the ``posix`` PHP
extension is required!

Change-Id: I8f3597c3b0856fc1159c7c173c0cfebcbf2278a6
Related: #49016
Releases: master