[FEATURE] Improve type converters
The ArrayConverter can now convert CSV strings (with configurable
delimiter) and JSON to array, depending on configuration. Conversion of
Resource instances to arrays is added with configurable file export
(base64 or external file).

The StringConverter accepts float, boolean, array and DateTime as source
types now. For DateTime conversion the W3C format is the default and can
be changed through configuration. Array sources can be converted to a
CSV string (with configurable delimiter)  or JSON, again depending on
configuration.

A ResourceConverter has been added. Unlike the ResourceTypeConverter it
is not for converting file uploads to Resource instances but handles
conversion of string and array sources.

Further changes:

* the BooleanConverter accepts integer and float as source types
* the IntegerConverter now can convert DateTime soures (to timestamp)
* add (more) class documentation for type converters to improve the
  converter reference in the Flow guide appendix

Change-Id: Ib5acdb59ee2ff9593456fe35960cdc9c41d71ace
Related: NEOS-158
Releases: master, 2.2