[BUGFIX] Respect correct property filling priority in ObjectConverter
The order preference to try to set a property
via constructor, via setter and via public property
is now respected correctly by the
``ObjectConverter::getTypeOfChildProperty()`` method.
That method used to check the setter annotation in the
first place, then the constructor annotation in the second
place, and failed when a property was only settable through
its public nature since that case was not backed at all.

The checking/setting priority now follows the one used by
``ObjectAccess``, so it is Constructor > Setter > Field.

Change-Id: I335294eed7d3699db88bf1ed64688c7f429a8889
Fixes: FLOW-33
Releases: master