[FEATURE] Allow logging of exceptions to be switched off
The AbstractExceptionHandler logs all exceptions before passing control
to the concrete ExceptionHandler. This means that even if an exception
is handled with a custom template (e.g. for exceptions grouped as 404
status errors) the error log is filled with exceptions and the traces
fill up the Data/Logs/Exceptions directory.

Now the new option ``logException`` can be used to switch that off. Flow
sets this to FALSE for the ``notFoundExceptions`` rendering group by
default.

Change-Id: Id6e523aeebcfdd366b7831f20c7a4634b494962b
Releases: master, 2.2