[FEATURE] Mark deprecated commands as such in help texts
This change adds a feature which detects @deprecated annotations
on command methods and marks those commands as deprecated
in the command help overview and help details for the respective
command. It also adds a warning line to the commands output
with a reference to a possible replacement command which
should be added to a @see annotation of the deprecated command.

Change-Id: Ie15273c282fa109780b862fc9fb09577890eca60
Releases: master