[FEATURE] The Query Object Model supports distinct queries
This changeset introduces the methods getDistinct and setDistinct on
the Query object to allow queries to explicitly return only distinct
result sets, which might be needed for join queries which happen
implicitly in subproperty queries, e.g.
property.subProperty.foo IN (1,2,3)

Doctrine automatically hydrates only distinct entities on result sets,
but that happens only after a limit clause on the query. This leads to
wrong query results with limit clauses, with less entities than
distinct existing entities in the database. A test is provided that shows
the behaviour.

Change-Id: I34ddbab23e760bfcb4be6f47992692adb6e20c61
Resolves: FLOW-21
Releases: master