[!!!][TASK] Introduce InjectSettings Annotation
This change mainly cleans up injection code by moving reading of
injection annotations to the ConfigurationBuilder from the
ProxyClassBuilder, so that the ProxyClassBuilder again mostly works
based off the given configuration.

Additionally property injection was moved to a separate annotation
that now allows injection of whole package settings from a separate
package with the following syntax::

  @Flow\InjectSettings(package="TYPO3.Party")

will inject all settings for the package ``TYPO3.Party``.

Just using the annotation like this::

  @Flow\InjectSettings

will inject all settings for the package in which the class with the
annotation is in.

Giving a specific setting path is also possible, with or without
the package. So this::

  @Flow\InjectSettings(package="TYPO3.Flow", path="i18n.defaultLocale")

Will inject the Setting TYPO3.Flow.i18n.defaultLocale regardless
in which class the annotation was used.

This change is marked breaking as injection of settings via
the ``Inject`` annotation is from now on deprecated and will be
removed in three versions.

Change-Id: Iaec291e40ffd352de9810c4e72027c455bf8c566
Releases: master
Resolves: FLOW-148