[!!!][FEATURE] Restructure policy component to new Policy.yaml format
This change introduces the new concept for policies and
privileges. It also includes a restructuring of the privilege
voting process.

This is a breaking change mainly because it drops support for content
security and secure downloads. Both features will be re-added by new
privilege types in separate changes.
Besides it is quite likely that custom code that interacts with the
(non-public) API of the security framework won't work without
adjustments.

The new ``Policy.yaml`` syntax is covered by code migrations, so make
sure to run::

  ./flow core:migrate
  ./flow doctrine:migrate

commands and to carefully read their output.

Change-Id: I84e188e89a05ec0dd1f9ee96fe312dac81806759
Migration: TYPO3.Flow-20141113121400
Resolves: FLOW-11
Releases: master