[BUGFIX] Call to undefined method ask in Cli CommandController
When required arguments are left out in a command controller call
an undefined method exception is thrown in Cli\CommandController.
This regression was introduced in Ia77c62b41fb598bdfb7b81c530494ba819a590d1

This change updates the call to the ask() method to use the ConsoleOutput
object.

Change-Id: I5961ce3a8f4accc8d27d67acd1214805e574de82
Fixes: FLOW-93
Releases: master