[TASK] Use unicode-safe pathinfo function and use it where necessary
``pathinfo()`` function is not unicode-friendly
if setlocale is not set. It's sufficient to set it
to any UTF-8 locale to correctly handle unicode strings.
This change temporarily sets locale to 'en_US.UTF-8'
and then restores original locale.
It's not necessary to use this function in cases,
where only file extension is determined, as it's
hard to imagine a unicode file extension.

Change-Id: Ia1df0375fa5b0926f0efe4881f9b8da69cd0cced
Related: FLOW-101
Releases: master, 1.2