[!!!][FEATURE] Multi-Storage / Multi-Target Resource Management
This change introduces a revised resource management which allows
for storage and publication of persistent or static resources (assets)
in the local file system or other services, such as Amazon S3 or
Rackspace CloudFiles. It also introduces the concept of collections
which allows for grouping resources into collections with specific
storage and publication rules.

Existing persistent resources are migrated through the Doctrine
migration contained in this feature.

Note: this change raises the PHP requirement to 5.5.

Resolves: FLOW-108
Releases: master
Change-Id: Ia2b47b4070a2dfabf4833bf1f0f3967ba3b032a7