[TASK] Move privilege evaluation into privilege manager
To avoid usage of a static vote functions in privilege
classes, this change moves evaluation of privileges into
the privilege manager.
This change removes the concept of privilege voters, which
is not needed due to the posssibility of implementing custom
privilege types.

This change also fixes an inconsistency within the privilege
evaluation process: Privilege targets with runtime evaluations
will no longer taken into account, if the runtime constraint
does not match the current situation.

When setting the same privilege twice within the same role,
only the last one will have effect. With that it is actually possible
to override permissions, e.g. in a dependant package.

Change-Id: Idc76f8f013893116c56db13f186d35a0842fc50d
Related: FLOW-11
Releases: master