[TASK] Make "renderingGroup" available to custom exception handlers
This change adjusts the exception handling slightly to make the
resolved "renderingGroup" available to custom exception handlers.

Besides, this sets the Fluid StandaloneView request package to
"TYPO3.Flow" for depending ViewHelpers to work properly.

This is currently required for TYPO3.Neos in order to localize the
error messages.

Change-Id: I3ac02f17f9aeba428bcd4d0b95b99b4dd7ed7d83
Related: NEOS-497
Releases: master