[BUGFIX] Constraint with "IN" and empty collection should work
Entity constraints using "IN" where the argument resulted in an
empty array would generate a query that contained an empty ``IN()``
operation. This breaks at least in MySQL. The only way to test for
NULL is a IS NULL constraint. This change takes care of that.
The changed test exposed the issue.

Change-Id: Ie5ef0e336699dcb7c5b5fdde03939d0e0b09ddb9
Releases: master