[BUGFIX] ValidatorResolver shouldn't ask class loader for simple types
The ValidatorResolver generates a Validator conjunction for a given class
if that class exists. Problem is that there is no check for simple types
before this, so a simple type will trigger the class loader (every time).
To avoid that we now check if the given type is a simple type.

Change-Id: I483666c78321534476655a841a5b0bb15d941da2
Releases: master, 3.0, 2.3, 2.2
Related: Icefc6eadbe7cae2e12cc2204ec6de3f25370aefa