[!!!][TASK] Remove obsolete "security.enable" Setting
This change removes the `TYPO3.Flow.security.enable` and all mentions
and usages of it.

Background:

This setting was initially intended for performance reasons (For
applications without security features) and in order to disable
security for (functional) tests.
For the latter we use a different approach since a while and the
performance hit of security features is also neglect-able since
Flow pre-compiles classes (at least if there is no complex policy
configured).
Besides the flag was never evaluated consistently.

Change-Id: I33be5efd2a23102d4c4957f671391cc12d8a047c
Resolves: FLOW-181
Related: FLOW-11
Releases: master