[TASK] Unclutter Storage API and keep upload internal
This removes the ``importUploadedResource()`` method from ``Collection``
and ``WriteableStorage`` to clean up the interface.
Uploaded files are now preprocessed in the ``ResourceManager`` and then handled
via the ``importResource()`` method. Additionally the ``$filename`` argument
is removed from ``Collection::importResourceFromContent()`` and
``Storage::importResourceFromContent()`` as it can be handled in the
``ResourceManager``.

Also includes fixes for environments with activated ``open_basedir`` directive.

Change-Id: Ifc4da1580dfb983e529380e18d2a0668cd90e92f
Releases: master