[BUGFIX] Interface object configuration has no effect
This fixes an issue with the object configuration for interfaces which
resulted in possibly defined configuration objects in Objects.yaml to
be ignored.

Even though Flow internally uses objects in dependency injection
referenced through the interface name (for example
"PackageManagerInterface"), the object configuration for such objects
did not actually work. The error only remained undiscovered because
the object configuration of the respective implementation classes had
meaningful options set.

With this change applied, it is now possible to safely inject an
"interface object".

Change-Id: I5e1e90f22881e7223632c10c274aac3e476edcbf
Resolves: FLOW-187
Releases: master