[!!!][BUGFIX] SessionManagerInterface and SessionInterface are incomplete
This change adds functions which have been around for some time now
in Session and SessionManager to their respective interfaces.

TransientSession now also implements these previously missing methods.

This patch is breaking in the unlikely case that you implemented your
own Session or SessionManager implementation and forgot to implement the
methods mentioned in the classes but not yet mentioned in the interfaces.

Change-Id: I18d439b830ac05fe5294af8199b1f71ff3bcf350
Releases: master