[FEATURE] Add command to set charset/collation on MySQL
This adds a new command to set the character set and collation used in
MySQL::

  ./flow database:setcharset

It will convert the database configured in the settings and all tables
inside to use a default character set of utf8 and a default collation
of utf8_unicode_ci. If needed, those defaults can be overridden.

It will also convert all character type columns to that combination of
charset and collation.

Change-Id: I60419d9db5cb3cf50d6c71316c443f400cdb2079
Releases: master
Related: NEOS-800