[FEATURE] Reconnect the EntityManager automatically
If the persistence backend closes the connection in the background.
E.g. if the MySQL server is configured with a low wait_timeout setting
of 10-30 seconds.
The ``PersistenceManager`` now catches the exception, establishes a new
connection automatically and tries to send all entity manager changes
to the backend again.
This is really important for long running commands or tasks.

Without this central improvement in the persistence manager it's really hard
to build stable daemon workers based on Flow commands.

Change-Id: I8e0fecbfe1bc63fa0aae24c0c049ebeaffc604cd
Releases: master
Resolves: FLOW-207