[TASK] Cleanup FileMonitor
This does a cleanup of the ``FileMonitor`` at the same time improving
performance. Reading the current state of directories is now done
with a ``Generator`` method to reduce memory footprint. Additionally
the amount of looping was reduced.

Finally you can now exclude whole subdirectories by adding an empty
file named ``.flowFileMonitorIgnore`` to it.

Change-Id: I76ee94277ab3b4512c8262123ab9401952786c49
Releases: master, 3.0